+++ Deosrc.MoviesTechnicalTest.Api/Entities/Movie.cs
@@ -6,9 +6,9 @@ namespace Deosrc.MoviesTechnicalTest.Api.Entities
         public required string Title { get; set; }
         public required DateOnly ReleaseDate { get; set; }
         public required string Overview { get; set; }
-        public required double Popularity { get; set; }
+        public required decimal Popularity { get; set; }
         public required int VoteCount { get; set; }
-        public required double VoteAverage { get; set; }
+        public required decimal VoteAverage { get; set; }
         public required string OriginalLanguage { get; set; }
         public required string Genre { get; set; }
         public required string PosterUrl { get; set; }
